requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
AnyQt==0.2.0
APScheduler==3.11.0
argcomplete==3.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
bcc==0.29.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
docopt==0.6.2
enum34==1.1.10
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
freetype-py==2.4.0
frozenlist==1.5.0
fuse-python==1.0.7
gpg==1.18.0
greenlet==3.0.3
httplib2==0.20.4
idna==3.6
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.3
multidict==6.1.0
mutagen==1.46.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
ntpsec==1.2.2
numpy==2.1.3
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pipx==1.4.3
platformdirs==4.2.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pylibacl==0.7.0
pynvim==0.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
pytz==2024.2
pyxattr==0.8.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.2.0
qq-botpy==1.2.1
referencing==0.35.1
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rlPyCairo==0.3.0
rpds-py==0.22.3
scv==2.0.0
setuptools==68.1.2
sip==6.8.6
six==1.16.0
soupsieve==2.6
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==235
tinycss2==1.4.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.2
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
virtualenv==20.28.0
wadllib==1.3.6
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
wheel==0.42.0
win2xcur==0.1.2
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.18.3
yt-dlp==2024.4.9
